# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Directorio de Cartago API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /businesses:
    get:
      operationId: getBusinesses
      tags: [businesses]
      summary: List businesses with optional filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Businesses list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessListResponse"
    post:
      operationId: createBusiness
      tags: [businesses]
      summary: Register a new business
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessRequest"
      responses:
        "201":
          description: Business created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /businesses/{id}:
    get:
      operationId: getBusiness
      tags: [businesses]
      summary: Get business details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateBusiness
      tags: [businesses]
      summary: Update business (owner or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessRequest"
      responses:
        "200":
          description: Business updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    delete:
      operationId: deleteBusiness
      tags: [businesses]
      summary: Delete business (owner or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /businesses/{id}/approve:
    post:
      operationId: approveBusiness
      tags: [businesses]
      summary: Approve a business (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /businesses/{id}/reject:
    post:
      operationId: rejectBusiness
      tags: [businesses]
      summary: Reject a business (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectRequest"
      responses:
        "200":
          description: Business rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
...
  /subscriptions/subscribe:
    post:
      operationId: subscribe
      tags: [subscriptions]
      summary: Subscribe to a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscribeRequest"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        businessCount:
          type: integer
      required: [id, name, businessCount]
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        description:
          type: string
      required: [name]
    BusinessImage:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        isPrimary:
          type: boolean
      required: [id, url, isPrimary]
    Business:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        address:
          type: string
        phone:
          type: string
          nullable: true
        whatsapp:
          type: string
          nullable: true
        instagram:
          type: string
          nullable: true
        facebook:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        googleMapsUrl:
          type: string
          nullable: true
        googleMapsEmbed:
          type: string
          nullable: true
        schedule:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        ownerId:
          type: integer
        ownerName:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        images:
          type: array
          items:
            $ref: "#/components/schemas/BusinessImage"
        averageRating:
          type: number
          nullable: true
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, address, ownerId, status, images, reviewCount, createdAt]
    CreateBusinessRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        address:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        instagram:
          type: string
        facebook:
          type: string
        website:
          type: string
        googleMapsUrl:
          type: string
        schedule:
          type: string
        categoryId:
          type: integer
        images:
          type: array
          items:
            type: string
      required: [name, address]
    BusinessListResponse:
      type: object
      properties:
        businesses:
          type: array
          items:
            $ref: "#/components/schemas/Business"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [businesses, total, page, totalPages]
    RejectRequest:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
...
    Subscription:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
          nullable: true
        planId:
          type: integer
        planName:
          type: string
          nullable: true
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        isActive:
          type: boolean
        paymentReference:
          type: string
          nullable: true
      required: [id, userId, planId, startDate, endDate, isActive]
    SubscribeRequest:
      type: object
      properties:
        planId:
          type: integer
        paymentReference:
          type: string
      required: [planId]
...