# lib/api-zod/src/generated/types/business.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Directorio de Cartago API
 * OpenAPI spec version: 0.1.0
 */
import type { BusinessImage } from "./businessImage";
import type { BusinessStatus } from "./businessStatus";

export interface Business {
  id: number;
  name: string;
  description?: string | null;
  address: string;
  phone?: string | null;
  whatsapp?: string | null;
  instagram?: string | null;
  facebook?: string | null;
  website?: string | null;
  googleMapsUrl?: string | null;
  googleMapsEmbed?: string | null;
  schedule?: string | null;
  categoryId?: number | null;
  categoryName?: string | null;
  ownerId: number;
  ownerName?: string | null;
  status: BusinessStatus;
  images: BusinessImage[];
  averageRating?: number | null;
  reviewCount: number;
  createdAt: Date;
}