# lib/api-zod/src/generated/types/businessImage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Directorio de Cartago API
 * OpenAPI spec version: 0.1.0
 */

export interface BusinessImage {
  id: number;
  url: string;
  isPrimary: boolean;
}