# lib/api-zod/src/generated/types/createBusinessRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Directorio de Cartago API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBusinessRequest {
  name: string;
  description?: string;
  address: string;
  phone?: string;
  whatsapp?: string;
  instagram?: string;
  facebook?: string;
  website?: string;
  googleMapsUrl?: string;
  schedule?: string;
  categoryId?: number;
  images?: string[];
}