# lib/api-zod/src/generated/types/subscribeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Directorio de Cartago API
 * OpenAPI spec version: 0.1.0
 */

export interface SubscribeRequest {
  planId: number;
  paymentReference?: string;
}